It compiles, still get warnings may fix them. Needs to implement opengl or vulkan for render

+++ linuxdoom-1.10/r_draw.c
@@ -30,6 +30,7 @@ rcsid[] = "$Id: r_draw.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
 #include "doomdef.h"
 
+#include <stdint.h>
 #include "i_system.h"
 #include "z_zone.h"
 #include "w_wad.h"
@@ -461,7 +462,7 @@ void R_InitTranslationTables (void)
     int		i;
 	
     translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+    translationtables = (byte *)(( (intptr_t)translationtables + 255 )& ~255);
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
